+++ regression/reference/enum-cxx/wrapenum.cpp
@@ -41,7 +41,7 @@ int ENU_convert_to_int(enum ENU_Color in)
 // ----------------------------------------
 // Argument:  enum Color in
 // Statement: f_in_native_scalar
-int ENU_convert_to_int_bufferify(int in)
+int ENU_convert_to_int_bufferify(short in)
 {
     // splicer begin function.convert_to_int_bufferify
     Color SHCXX_in = static_cast<Color>(in);
